backend/src/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.1.0
autopep8==1.5.4
beautifulsoup4==4.9.1
black==19.10b0
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1
Django==3.1.1
idna==2.10
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.7.0
pip==20.2.2
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
setuptools==49.6.0.post20200814
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wheel==0.35.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
